Update data node management with example of moving a chunk
The data node management section now has a basic description of moving
a chunk between data nodes. This complements the section on adding new
data nodes (elasticity) with information on how to actually get data
to a new node.

The information on copying and moving chunks is now split across the
data node management section and the native replication section. In
the future, we might want to rewrite this to have everything about
copying and moving chunks in a single place.