Improve the multi-node HA section
Make it clear in the multi-node HA section that there are two viable
ways to provide multi-node HA. In particular, we should emphasize that
native replication is still not recommended for production use and
suggest setting up standby nodes as an alternative.

The change also fixes a couple of issues like text that ends in the
middle of the sentence, missing content that existed in old docs, etc.

Co-authored-by: Lana Brindley <github@lanabrindley.com>