Fix apt installation instructions (#710)
* Fix apt installation instructions

While individual packages in the apt repository are not signed the package list still is though. So we need to install our GPG key for package list verification. This PR adds this missing command and also adds lsb-release,  apt-transport-https and wget to the list of packages to install cause they are used in the instructions but might not be installed by default on some installations.

* Update installation-debian.md

Co-authored-by: Jacob Prall <prall.jacob@gmail.com>