Update page-index.js
Not sure how I've missed this outline previously, but we want people to save space with compression on raw data before we drop old data. This probably came from earlier discussions that a lot of stock/crypto folks seem to want to drop raw data ASAP, so totally understandable.

With this order (and assuming data we provide is ~1 month of data), users get raw data in -> continuous aggregates for rollup and faster query -> compress data that's older than 2 weeks -> drop raw data that's older than 3.

This feels like a better overall flow.