tdigest and uddsketch links were swapped (#1137)
Also removed backticks in the links that were causing 404 errors.

Edited a paragraph in uddsketch that had a "the the" typo and a
sentence that was a little confusing.

Edited a tdigest paragraph that was using 'they' instead of 'it' to
refer to the aggregate function. That seemed unnatural to me, but I
can revert it if it was intentional.