[Vale] Improve link and semicolon linting
Markdown links:

Markdown link checking stopped working. (I suspect a breaking change in
Vale, because I'm positive this used to work.) This fixes that issue.

Deprecated links:

This tests for link references that aren't in the main body of the text
anymore, so we don't have so many dead link references at the bottom of
our documents.

Semicolons:

This ignores (in a somewhat hacky way) semicolons in import statements.

MDX:

Also lint MDX files.