Fix tag name for Docker image (#1371)
Somewhat confusingly, the TimescaleDB Docker image is tagged
'latest-pg14` but the TimescaleDB **HA** Docker image is tagged
`pg14-latest`. The documentation had these mixed up in some places.

Also changed the extended instructions to use the HA image rather than
the non-HA image. That is now the recommended image as it includes
Toolkit.