Rewrite data model section (#1401)
* Rewrite data model section

The data model section is a little confusing. Readers have asked if a
narrow-table model requires a separate table for each metric, which is
not the case. Rewrote to add a specific example of how it looks in
TimescaleDB, to emphasize there is one table but many (short) rows.

Also, some of the explanation had room for tightening up to focus on the
main points and make some examples more concrete.

Finally, recombined everything from three files into one. As a
learning piece, each section does depend slightly on the others, and it
was too fragmented before. Tried to write them independently enough that
you _can_ skip to a particular header if you wish, but it is really one
argument, and not too long for a single piece.

Co-authored-by: Lana Brindley <github@lanabrindley.com>