Clean up API frontmatter
Clean up API frontmatter. Namespace API- and hyperfunction-specific
fields so things are easier to find.

Convert `topic` (singular) to `topics` (plural) to accommodate APIs that
relate to multiple topics (for example, both compression and jobs).

Mark fields that will be deprecated (as soon as the hyperfunction tables
component is updated).

Convert the hyperfunction tables in `About hyperfunctions` to be sorted
by type:
*   Makes it easier to find things and makes the structure of families
    clearer
*   There is currently a bug with the non-sorted tables