[Decompression] Update SQL for finding job ID (#1455)
* [Decompression] Update SQL for finding job ID

The SQL query given for finding job ID was updated. There is an easier
way to do it now, without an unnecessary `JOIN`.

* Rearrange decompress and backfilling sections

While converting backfilling instructions to a procedure, noticed a lot
of repeated sections between these two pages. Backfilling was already
documented (twice on two different pages). Removed the repetition and
cleaned up a few things to be more in line with our current style.

* Minor cleanup

Co-authored-by: Lana Brindley <github@lanabrindley.com>