[Cloud] Integrations for telemetry exports (#1367)
* [Cloud] Integrations for telemetry exports

Integrations for telemetry exports to DataDog and AWS CloudWatch.

* Rearrange from processes to procedures

Rearrange so creating and attaching exporters is documented as processes rather than procedures.

* Specify which metrics are exported

* Correct and minor improvements

* Add warning about securing AWS keys

* Improve description of IAM role field

* Add info about where to view metrics

Add a few lines about where to go see the metrics in Datadog and
CloudWatch.

* Add variables to includes

Use variables to add the integration name to the partials. This creates
unique IDs for the partial headings so we can link to them from the
process steps.