Improve HyperLogLog documentation (#1590)
* Emphasize warning about low bucket count

Using a value for buckets that's less than 1024 results in very poor
accuracy, and is almost never wanted. As such, emphasize the warning
about this pitfall.

* Remove outdated information about HyperLogLog

This used to be true, but since this was written Toolkit has been
updated to use HyperLogLog++, which has better accuarcy for low
cardinality datasets.