Move redirects to docs repo
Redirects don't actually need to sit in the web-documentation repo,
because they're copied and pasted elsewhere. It will be easier to keep
them up to date if we store them alongside the content, where all the
link and URL changing happens. That means we can update the list
alongside any PR that will break a link.

Also took this opportunity to reorganize by alphabetizing the list and
removing duplicates.

Finally tried to fix the `num_vals` redirect, which currently loops
until it hits the redirect limit.