Minor fixes to display of cagg APIs
It has been pointed out that the cagg API splash page conveys no useful
information and is actually kind of confusing because it directs away
from the API section.

As a short-term fix, remove the splash page and redirect to the first
child page. Add the link to the how-to guides to that child page.

Also fixes inconsistent header levels across the entire cagg API
section.