Expectations on the maintenance behavior with a replica. (#2755)
Correct the description of the maintenance behavior when replicas are present. In the previous version, we promisted our customers no downtime and two failover events, while in reality, the maintenance may result in up to two write downtimes, up to one minute each since the replica waits for up to a minute for the primary to restart.