Use schema qualified function
The documentation for migration mentions the use of
"timescaledb_post_restore" function post the loading of the data.

However, the standard dump file generated using pg_dump unsets the
search_path. So, it's important to use the schema qualified function.
Otherwise the command errors out at the end because it cannot find the
"timescaledb_post_restore" function.