Explain how encryption works in cloud (#2959)
* Explain how encryption works in cloud

We used a strong statement "All connections to Timescale are
encrypted.", which is not correct. In fact customers can connect to
their databases without using TLS and we need to address this in
documentation. It's yet to be defined if we want to make this user
configurable, but for now we need to update the doc to reflect the
reality.

* Update strict-ssl.md

Signed-off-by: Ana Tavares <110231687+contentana@users.noreply.github.com>

---------

Signed-off-by: Ana Tavares <110231687+contentana@users.noreply.github.com>
Co-authored-by: Ana Tavares <110231687+contentana@users.noreply.github.com>